[contacts._full_color_list] splice orphan fragments
This change's only purpose is not to break the logic of the method
   one residue -> one assigned color,
regardless of whether any particular residue gets plotted or not. This
way, the downstream color-assigner if colors is None gets a set of
fragment definitions that cover all residues.

Otherwise, the color-guesser will complain about not having information
for all residues.

Could have side effects but this private method is pretty controlled
by the methods around it, s.t. the residues ultimately passed to e.g.
a flareplot will know to discard anything that was in the orphan fragments