[utils.sequence.my_bioalign] Update to Biopython's new aligment method (updated one test as well)
For a quick patch release, and reproducible CI tests, I've tried to change as little as possible
outside the method itself, making the new method return alignment objects that behave like
the old alignment objects that the rest of mdciao is expecting. Mid-term, the intention
is to also change the rest of mdciao to work better with the new Align method of biopython.