[REAMDE.rst] workaround pypa's change that doesn't allow inclusion of images
The workaround is just to comment out the inclusion of files (images).

This is only for the purpose of the release, i.e. the pypi and the github release should be identical

Shortly after releasing this commit is reverted, s.t. top-level README.rst gets rendered with images on github.